Make the tests a little bit more robust
Previously, `done` could be called multiple times per test (in theory,
at least). Moreover, the `done` callback would be called before the
client was fully torn down. Both of these things would cause unrelated
tests to fail in "impossible" ways.

This change makes the tests always try very hard to call `done` exactly
once in the channel 0 callback, and if it's not possible, do that in a
place that guarantees that the client won't continue doing stuff. Since
now this requires all of the changed tests to clean up after themselves
and in some cases actually hear back from goval, the timeout has been
bumped a bit to account for that.

Also, all the `setUnrecoverableErrorHandler()` calls that were not
handling test logic were removed. Tests will fail _loudly_ due to the
thrown exception.