Relax the criteria to consider the WebSockets implementation as "working" (#94)
Previously, it was required to receive a valid api.Command over the WS
connection to deem the WebSockets implementation as "working". When
there is an incident, some of the WS connections are rejected early and
clients will unnecessarily fall back to using the polling
implementation, causing unnecessary churn.

This change makes the criteria to consider the websockets as working to
be just the fact that the `onopen` event was fired.