Reland: "Allow recycling of connection metadata upon reconnects (#96)"
Getting new tokens every time we want to reconnect is slightly wasteful,
since most likely the token the clients have already are still valid.
Furthermore, during incidents where goval is unhappy, all the extra
retries will cause cascading failures due to all the attempts to re-get
tokens, which causes extra load to repl-it-web as well as Lore.

This change now keeps the connection metadata across reconnects,
_unless_ goval explicitly tells crosis to fetch a new one. This happens
when goval accepts the connection and immediately closes it with the
Close Code 1008 (Policy Violation).

This time around, the `ConnectOptions` now receives a
`reuseConnectionMetadata` field, and will reuse tokens only when it's
set to `true`. This allows a flagged deployment.