Provide channel service / name on the crosis logs (#119)
Previously one had to pay attention and manually correlate what channels
correspond to what services when inspecting the crosis logs. That's a
bit too hard for humans, and the machines can do that for us.

This change adds the channel service (if available) and name (if
available) for all crosis logs.