We largely agreed not to do this, or to do it by removing FetchConnectionMetadataError.Aborted entirely, but in case we want the discussion.
These two signals currently have to match (user and API), which means the only way a client is allowed to abort is if the abortController said so. With that in mind, it doesn't make a lot of sense to me to require the client to return aborted if we're going to bail anyway.

This is a change that implements that thought. To ship, I'd rather just delete the concept of the client returning aborted entirely and just have its return value as meaningless if we're aborted.